[*] перенос функции
