+++ sources/libs/translation_module/include/translation_module/fillers.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <memory>
 #include <json/json.h>
 #include "aliases.h"
 
@@ -47,9 +48,37 @@ private:
     StringMap &stringMap;
 };
 
-std::unique_ptr<IFiller> createIFiller()
-{
 
+std::unique_ptr<IFiller> createFiller(std::string value, JsonMap &jsonMap, StringMap &stringMap) {
+    auto eraseBorders = [](std::string &str) {
+        if (str.size() >= 2) {
+            str = std::string(str.begin() + 1, str.end() - 1);
+        } else {
+            throw std::runtime_error("string is too small");
+        }
+    };
+
+    auto isBorderedBy = [](std::string &str, char b, char e) {
+        if (str.size() < 2) {
+            return false;
+        }
+        return str.front() == b && str.back() == e;
+    };
+
+    if (isBorderedBy(value, '{', '}')) {
+        eraseBorders(value);
+        return std::unique_ptr<IFiller>(new JsonFiller(value, jsonMap));
+    }
+
+    if (isBorderedBy(value, '"', '"')) {
+        eraseBorders(value);
+        if (isBorderedBy(value, '<', '>')) {
+            eraseBorders(value);
+            return std::unique_ptr<IFiller>(new StringFiller(value, stringMap));
+        }
+    }
+
+    throw std::runtime_error("Некорректное значение строки");
 }
 
 }
